requirements.txt
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
email-validator==2.2.0
flask==3.1.1
flask-sqlalchemy==3.1.1
flask-socketio==5.3.6
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
python-socketio==5.10.0
sqlalchemy==2.0.41
typing-extensions==4.14.1
werkzeug==3.1.3
flask-login==0.6.3
python-dotenv==1.0.1
oauthlib==3.3.1
certifi==2025.7.14
charset-normalizer==3.4.2
requests==2.32.4
urllib3==2.5.0